feat: add test for metrics federate endpoint
Added a test to ensure that the expected user facing Kafka metrics are returned from the
KAS Fleet Manager /kafkas/{id}/metrics/federate endpoint. Response should be in a Prometheus Text format.

A waitFor is used in the checks as metrics do not become available instantly as soon as the Kafka status becomes
ready.