Assert that after ownership is changed we delete orphan ACL bindings from kafka
Why:
We are working to improve the ACL management user experience. One problem is that after
a change of ownership, if the new user had any ACL bindings stored in the kafka control
plane, those ACL bindings were visible when listing ACLs. But this is misleading because
once they are the owner their authorization is entirely handled by the kas-broker-plugin
and will not match what was stored in kafka. So we have added code to delete these orphaned
ACL bindings when starting up the broker.

The test checks that a binding is deleted for the new owner, and also that we do not delete
some other principal's bindings.